Refactor crazyswarm cpp
This major refactor:
- separates the original implementation into proper header/source files
- lifts implementations of CrazyflieROS and CrazyflieServer into a
  separate library
- adds missing `rcl_interfaces` dependency to package
- reformats code with google C++ format style; add clang format file
