+++ .clang-format
@@ -1,0 +1,5 @@
+BasedOnStyle: Google
+IndentWidth: 4
+ColumnLimit: 120
+AccessModifierOffset: -4
+EmptyLineBeforeAccessModifier: Always
