+++ .gitignore
@@ -3,3 +3,4 @@ crazyswarm2_sim_webots/
 *.pyc
 *.o
 *~
+*.vscode/
